+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0-alpha33] - 2019-11-15
+### Add
+ - Added factory unit tests.
+ - Added travis yaml file.
+### Fix
+ - Fixed Condition factory rendering function.
+
 ## [1.0.0-alpha32] - 2019-11-14
 ### Fix
  - Text factory: